Use compiletest to make sure unsafe things do not compile
