+++ Cargo.toml
@@ -29,3 +29,4 @@ gcc = { version = "0.3.52", optional = true }
 
 [dev-dependencies]
 rustyline = "1.0.0"
+compiletest_rs = "0.3.0"
